Remove unused legacy components and theming
